sign-language-detector-python/requirements.txt
flask
flask-cors
numpy
pillow
mediapipe
scikit-learn
